+++ ntex/src/http/client/h1proto.rs
@@ -176,7 +176,7 @@ impl Stream for PlStream {
                     Err(RecvError::KeepAlive) => {
                         Err(io::Error::new(io::ErrorKind::Other, "Keep-alive").into())
                     }
-                    Err(RecvError::StopDispatcher) => {
+                    Err(RecvError::Stop) => {
                         Err(io::Error::new(io::ErrorKind::Other, "Dispatcher stopped")
                             .into())
                     }