+++ ntex/tests/connect.rs
@@ -8,21 +8,49 @@ use ntex::service::{fn_service, Service, ServiceFactory};
 use ntex::util::Bytes;
 
 #[cfg(feature = "openssl")]
+fn ssl_acceptor() -> tls_openssl::ssl::SslAcceptor {
+    use tls_openssl::ssl::{SslAcceptor, SslFiletype, SslMethod};
+
+    // load ssl keys
+    let mut builder = SslAcceptor::mozilla_intermediate(SslMethod::tls()).unwrap();
+    builder
+        .set_private_key_file("./tests/key.pem", SslFiletype::PEM)
+        .unwrap();
+    builder
+        .set_certificate_chain_file("./tests/cert.pem")
+        .unwrap();
+    builder.build()
+}
+
+#[cfg(feature = "openssl")]
 #[ntex::test]
-async fn test_string() {
+async fn test_openssl_string() {
+    use ntex::server::openssl;
+    use tls_openssl::ssl::{SslConnector, SslMethod, SslVerifyMode};
+
     let srv = test_server(|| {
-        fn_service(|io: Io| async move {
+        ntex::pipeline_factory(fn_service(|io: Io<_>| async move {
+            let res = io.read_ready().await;
+            assert!(res.is_ok());
+            Ok(io)
+        }))
+        .and_then(openssl::Acceptor::new(ssl_acceptor()))
+        .and_then(fn_service(|io: Io<_>| async move {
             io.send(Bytes::from_static(b"test"), &BytesCodec)
                 .await
                 .unwrap();
-            Ok::<_, io::Error>(())
-        })
+            Ok::<_, Box<dyn std::error::Error>>(())
+        }))
     });
 
-    let conn = ntex::connect::Connector::default();
-    let addr = format!("localhost:{}", srv.addr().port());
+    let mut builder = SslConnector::builder(SslMethod::tls()).unwrap();
+    builder.set_verify(SslVerifyMode::NONE);
+
+    let conn = ntex::connect::openssl::Connector::new(builder.build());
+    let addr = format!("127.0.0.1:{}", srv.addr().port());
     let con = conn.call(addr.into()).await.unwrap();
-    assert_eq!(con.query::<PeerAddr>().get().unwrap(), srv.addr().into());
+    let item = con.recv(&BytesCodec).await.unwrap().unwrap();
+    assert_eq!(item, Bytes::from_static(b"test"));
 }
 
 #[cfg(feature = "rustls")]